backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn[standard]==0.24.0
pytesseract==0.3.10
pypdf==3.17.0
pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
chromadb==0.5.3
sentence-transformers==2.7.0
neo4j==5.25.0